Updated entity system

+++ src/snail_jumpy_intrinsics.h
@@ -1,6 +1,5 @@
 #if !defined(SNAIL_JUMPY_INTRINSICS_H)
 #define SNAIL_JUMPY_INTRINSICS_H
-
 typedef struct _bit_scan_result bit_scan_result;
 struct _bit_scan_result
 {
@@ -13,10 +12,9 @@ struct _bit_scan_result
 #pragma intrinsic(_BitScanForward)
 
 internal inline bit_scan_result
-ScanForLeastSignificantSetBit(u32 Mask)
-{
+ScanForLeastSignificantSetBit(u32 Mask){
     bit_scan_result Result;
-    Result.Found = _BitScanForward64(&(DWORD)Result.Index, Mask);
+    Result.Found = _BitScanForward64(&(unsigned long)Result.Index, Mask);
     return(Result);
 }
 
@@ -24,8 +22,7 @@ ScanForLeastSignificantSetBit(u32 Mask)
 #include <intrin.h>
 
 internal inline bit_scan_result
-ScanForLeastSignificantSetBit(u32 Mask)
-{
+ScanForLeastSignificantSetBit(u32 Mask){
     bit_scan_result Result;
     Result.Found = _BitScanForward64((unsigned long *)&Result.Index, Mask);
     return(Result);
